thread

+++ internal/util-logging/src/main/scala/sbt/internal/util/ConsoleAppender.scala
@@ -643,40 +643,50 @@ private[sbt] object ProgressState {
    * at the info or greater level, we can decrement the padding because the console
    * line will have filled in the blank line.
    */
-  private[sbt] def updateProgressState(pe: ProgressEvent, terminal: Terminal): Unit = {
+  private[sbt] def updateProgressState(
+      pe: ProgressEvent,
+      terminal: Terminal
+  ): Unit = {
     val state = terminal.progressState
     val isRunning = terminal.prompt == Prompt.Running
     pe.command.foreach(state.currentCommand.set)
     if (terminal.isSupershellEnabled) {
-      terminal.withPrintStream { ps =>
-        val info = if (isRunning) {
-          pe.items.map { item =>
-            val elapsed = item.elapsedMicros / 1000000L
-            s"  | => ${item.name} ${elapsed}s"
+      if (!pe.skipIfActive.getOrElse(false) || !isRunning) {
+        terminal.withPrintStream { ps =>
+          val info = if (isRunning) {
+            pe.items.map { item =>
+              val elapsed = item.elapsedMicros / 1000000L
+              s"  | => ${item.name} ${elapsed}s"
+            }
+          } else {
+            pe.command.toSeq.map { cmd =>
+              s"sbt server is running `$cmd`. Type `kill $cmd` to stop evaluation."
+            }
           }
-        } else {
-          pe.command.toSeq.map { cmd =>
-            s"sbt server is running `$cmd`. Type `kill $cmd` to stop evaluation."
+
+          val currentLength = info.foldLeft(0)(_ + terminal.lineCount(_))
+          val previousLines = state.progressLines.getAndSet(info)
+          val prevLength = previousLines.foldLeft(0)(_ + terminal.lineCount(_))
+          val (height, width) = terminal.prompt match {
+            case Prompt.Running =>
+              terminal.getLastLine.map(terminal.getLineHeightAndWidth).getOrElse((0, 0))
+            case a => terminal.getLineHeightAndWidth(a.render())
           }
-        }
+          val prevSize = prevLength + state.padding.get
 
-        val currentLength = info.foldLeft(0)(_ + terminal.lineCount(_))
-        val previousLines = state.progressLines.getAndSet(info)
-        val prevLength = previousLines.foldLeft(0)(_ + terminal.lineCount(_))
-        val (height, width) = terminal.prompt match {
-          case Prompt.Running =>
-            terminal.getLastLine.map(terminal.getLineHeightAndWidth).getOrElse((0, 0))
-          case a => terminal.getLineHeightAndWidth(a.render())
+          val newPadding = math.max(0, prevSize - currentLength)
+          state.padding.set(newPadding)
+          state.printPrompt(terminal, ps)
+          ps.print(state.printProgress(height, width))
+          ps.flush()
         }
-        val prevSize = prevLength + state.padding.get
-
-        val newPadding = math.max(0, prevSize - currentLength)
-        state.padding.set(newPadding)
-        state.printPrompt(terminal, ps)
-        ps.print(state.printProgress(height, width))
-        ps.flush()
+      } else if (state.progressLines.get.nonEmpty) {
+        state.progressLines.set(Nil)
+        terminal.withPrintStream { ps =>
+          ps.print(ClearScreenFromCursorToBottom)
+          ps.flush()
+        }
       }
     }
   }
-
 }
